Закончил реализацию интерфейсов
